HashCounter

hash with counter, to store requests, and count each time it is used

See also

Public fields

hash

(list) a list for internal use only, with elements key, sig, and count

Methods

Public methods

HashCounter$put()

HashCounter$get()

HashCounter$clone()

Method put()

Register a request by it's key

Usage

HashCounter$put(req_sig)

Arguments

req_sig

an object of class RequestSignature

Returns

nothing returned; registers request and iterates internal counter

Method get()

Get a request by key

Usage

HashCounter$get(req_sig)

Arguments

req_sig

an object of class RequestSignature

Returns

(integer) the count of how many times the request has been made

Method clone()

The objects of this class are cloneable with this method.

Usage

HashCounter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- HashCounter$new()
x$hash

#> list()

z <- RequestSignature$new(method = "get", uri = "https:/httpbin.org/get")
x$put(z)
x$hash

#> $`GET:  https:/httpbin.org/get`
#> $`GET:  https:/httpbin.org/get`$key
#> [1] "GET:  https:/httpbin.org/get"
#> 
#> $`GET:  https:/httpbin.org/get`$sig
#> <RequestSignature> 
#>   method: GET
#>   uri: https:/httpbin.org/get
#> 
#> $`GET:  https:/httpbin.org/get`$count
#> [1] 1
#> 
#> 

x$get(z)

#> [1] 1

x$put(z)
x$get(z)

#> [1] 2