StubCounter

hash with counter to store requests and count number of requests made against the stub

Public fields

hash

(list) a list for internal use only, with elements key, sig, and count

Methods

Public methods

StubCounter$put()

StubCounter$count()

StubCounter$clone()

Method put()

Register a request by it's key

Usage

StubCounter$put(x)

Arguments

x

an object of class RequestSignature

Returns

nothing returned; registers request & iterates internal counter

Method count()

Get the count of number of times any matching request has been made against this stub

Usage

StubCounter$count()

Method clone()

The objects of this class are cloneable with this method.

Usage

StubCounter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- StubCounter$new()
x
#> <StubCounter>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     count: function () 
#>     hash: list
#>     put: function (x) 
#>   Private:
#>     total: 0
x$hash
#> list()
x$count()
#> [1] 0
z <- RequestSignature$new(method = "get", uri = "https:/httpbin.org/get")
x$put(z)
x$count()
#> [1] 1
x$put(z)
x$count()
#> [1] 2