UriPattern

UriPattern

Arguments

pattern
(character) a uri, as a character string. if scheme is missing, it is added (we assume http)

regex_pattern
(character) a uri as a regex character string, see base::regex. if scheme is missing, it is added (we assume http)

Details

Methods

add_query_params

Add query parameters to the URI - query_params

matches(uri)

Match a uri against that given in pattern - uri (character) a uri, including scheme (i.e., http or https)

Examples

# trailing slash
(z <- UriPattern$new(pattern = "http://foobar.com"))

#> <UriPattern>
#>   Public:
#>     add_query_params: function (query_params) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (pattern = NULL, regex_pattern = NULL) 
#>     matches: function (uri) 
#>     pattern: http://foobar.com
#>     query_params: NULL
#>     regex: FALSE
#>     to_s: function () 

z$matches("http://foobar.com")

#> [1] TRUE

z$matches("http://foobar.com/")

#> [1] TRUE

# default ports
(z <- UriPattern$new(pattern = "http://foobar.com"))

#> <UriPattern>
#>   Public:
#>     add_query_params: function (query_params) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (pattern = NULL, regex_pattern = NULL) 
#>     matches: function (uri) 
#>     pattern: http://foobar.com
#>     query_params: NULL
#>     regex: FALSE
#>     to_s: function () 

z$matches("http://foobar.com:80")

#> [1] TRUE

z$matches("http://foobar.com:80/")

#> [1] TRUE

z$matches("http://foobar.com:443")

#> [1] TRUE

z$matches("http://foobar.com:443/")

#> [1] TRUE

# user info
(z <- UriPattern$new(pattern = "http://foobar.com"))

#> <UriPattern>
#>   Public:
#>     add_query_params: function (query_params) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (pattern = NULL, regex_pattern = NULL) 
#>     matches: function (uri) 
#>     pattern: http://foobar.com
#>     query_params: NULL
#>     regex: FALSE
#>     to_s: function () 

z$matches("http://user:pass@foobar.com")

#> [1] FALSE

# regex
(z <- UriPattern$new(regex_pattern = ".+ample\\.."))

#> <UriPattern>
#>   Public:
#>     add_query_params: function (query_params) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (pattern = NULL, regex_pattern = NULL) 
#>     matches: function (uri) 
#>     pattern: https?://.+ample\..
#>     query_params: NULL
#>     regex: TRUE
#>     to_s: function () 

z$matches("http://sample.org")

#> [1] TRUE

z$matches("http://example.com")

#> [1] TRUE

z$matches("http://tramples.net")

#> [1] FALSE

# add query parameters
(z <- UriPattern$new(pattern = "http://foobar.com"))

#> <UriPattern>
#>   Public:
#>     add_query_params: function (query_params) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (pattern = NULL, regex_pattern = NULL) 
#>     matches: function (uri) 
#>     pattern: http://foobar.com
#>     query_params: NULL
#>     regex: FALSE
#>     to_s: function () 

z$add_query_params(list(pizza = "cheese", cheese = "cheddar"))
z$pattern

#> [1] "http://foobar.com?pizza=cheese&cheese=cheddar"

(z <- UriPattern$new(pattern = "http://foobar.com"))

#> <UriPattern>
#>   Public:
#>     add_query_params: function (query_params) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (pattern = NULL, regex_pattern = NULL) 
#>     matches: function (uri) 
#>     pattern: http://foobar.com
#>     query_params: NULL
#>     regex: FALSE
#>     to_s: function () 

z$pattern

#> [1] "http://foobar.com"

z$add_query_params(list(pizza = "deep dish", cheese = "cheddar"))
z$pattern

#> [1] "http://foobar.com?pizza=deep%20dish&cheese=cheddar"

# any pattern
(z <- UriPattern$new(regex_pattern = ".+"))

#> <UriPattern>
#>   Public:
#>     add_query_params: function (query_params) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (pattern = NULL, regex_pattern = NULL) 
#>     matches: function (uri) 
#>     pattern: https?://.+
#>     query_params: NULL
#>     regex: TRUE
#>     to_s: function () 

z$regex

#> [1] TRUE

z$pattern

#> [1] "https?://.+"

z$matches("http://stuff.com")

#> [1] TRUE

z$matches("https://stuff.com")

#> [1] TRUE

z$matches("https://stuff.com/stff")

#> [1] TRUE

z$matches("https://stuff.com/apple?bears=3")

#> [1] TRUE